Split session results bundle
