+++ src/main/java/teammates/ui/output/UserSessionResultsData.java
@@ -1,0 +1,233 @@
+package teammates.ui.output;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+
+import teammates.common.datatransfer.SessionResultsBundle;
+import teammates.common.datatransfer.questions.FeedbackQuestionDetails;
+import teammates.common.util.Const;
+import teammates.storage.entity.FeedbackQuestion;
+import teammates.storage.entity.FeedbackResponse;
+import teammates.storage.entity.ResponseGiver;
+import teammates.storage.entity.ResponseInstructorComment;
+import teammates.storage.entity.ResponseRecipient;
+import teammates.storage.entity.Student;
+import teammates.storage.entity.User;
+
+/**
+ * API output format for user-specific session results, including statistics.
+ */
+public class UserSessionResultsData implements ApiOutput {
+
+    private static final String REGEX_ANONYMOUS_PARTICIPANT_HASH = "[0-9]{1,10}";
+
+    final List<UserQuestionOutput> questions = new ArrayList<>();
+
+    UserSessionResultsData() {
+        // use factory method instead
+    }
+
+    /**
+     * Factory method to construct API output for a specific user.
+     */
+    public static UserSessionResultsData initForUser(SessionResultsBundle bundle, User user) {
+        UserSessionResultsData sessionResultsData = new UserSessionResultsData();
+
+        Map<FeedbackQuestion, List<FeedbackResponse>> questionsWithResponses =
+                bundle.getQuestionResponseMap();
+
+        questionsWithResponses.forEach((question, responses) -> {
+            FeedbackQuestionDetails questionDetails = question.getQuestionDetailsCopy();
+            boolean hasResponseButNotVisibleForPreview = bundle.getQuestionsNotVisibleForPreviewSet()
+                    .contains(question);
+            boolean hasCommentNotVisibleForPreview = bundle.getQuestionsWithCommentNotVisibleForPreviewSet()
+                    .contains(question);
+
+            String questionStatistics = hasResponseButNotVisibleForPreview
+                    ? ""
+                    : questionDetails.getQuestionResultStatisticsJson(question, user.getId(), bundle);
+            UserQuestionOutput qnOutput = new UserQuestionOutput(question,
+                    questionStatistics,
+                    hasResponseButNotVisibleForPreview,
+                    hasCommentNotVisibleForPreview);
+            Map<ResponseRecipient, List<ResponseOutput>> otherResponsesMap = new HashMap<>();
+
+            qnOutput.getFeedbackQuestion().hideInformationForStudent();
+
+            if (!hasResponseButNotVisibleForPreview && questionDetails.isIndividualResponsesShownToStudents()) {
+                for (FeedbackResponse response : responses) {
+                    boolean isUserGiver = Objects.equals(user, response.getGiver().getGiverUser());
+                    boolean isUserRecipient = Objects.equals(user, response.getRecipient().getRecipientUser());
+
+                    ResponseOutput responseOutput = buildSingleResponseForUser(response, bundle, user);
+
+                    if (isUserRecipient) {
+                        qnOutput.responsesToSelf.add(responseOutput);
+                    }
+
+                    if (isUserGiver) {
+                        qnOutput.responsesFromSelf.add(responseOutput);
+                    }
+
+                    if (!isUserRecipient && !isUserGiver) {
+                        otherResponsesMap.computeIfAbsent(response.getRecipient(), k -> new ArrayList<>())
+                                .add(responseOutput);
+                    }
+
+                    qnOutput.allResponses.add(responseOutput);
+                }
+            }
+            qnOutput.otherResponses.addAll(otherResponsesMap.values());
+
+            sessionResultsData.questions.add(qnOutput);
+        });
+
+        return sessionResultsData;
+    }
+
+    private static ResponseOutput buildSingleResponseForUser(
+            FeedbackResponse response, SessionResultsBundle bundle, User user) {
+        Objects.requireNonNull(user);
+
+        ResponseGiver giver = response.getGiver();
+        boolean isUserGiver = Objects.equals(user, giver.getGiverUser());
+        boolean isUserTeamGiver = false;
+        if (user instanceof Student student) {
+            isUserTeamGiver = Objects.equals(student.getTeam(), giver.getGiverTeam());
+        }
+
+        String giverName;
+        String giverTeam = "";
+        if (isUserTeamGiver) {
+            giverName = String.format("Your Team (%s)", giver.getTeamName());
+            giverTeam = giver.getTeamName();
+        } else if (isUserGiver) {
+            giverName = "You";
+            giverTeam = giver.getTeamName();
+        } else {
+            giverName = removeAnonymousHash(
+                    SessionResultsData.getGiverNameOfResponse(response.getId(), giver, bundle));
+        }
+
+        ResponseRecipient recipient = response.getRecipient();
+        boolean isUserRecipient = Objects.equals(user, recipient.getRecipientUser());
+        boolean isUserTeamRecipient = false;
+        boolean isRecipientVisible = bundle.isResponseRecipientVisible(response.getId(), recipient.getRecipientType());
+        if (user instanceof Student student) {
+            isUserTeamRecipient = Objects.equals(student.getTeam(), recipient.getRecipientTeam());
+        }
+
+        String recipientName;
+        String recipientTeam = "";
+        if (isUserRecipient) {
+            recipientName = "You";
+            recipientTeam = recipient.getTeamName();
+        } else if (isUserTeamRecipient) {
+            recipientName = String.format("Your Team (%s)", recipient.getDisplayName());
+            recipientTeam = recipient.getTeamName();
+        } else {
+            recipientName = removeAnonymousHash(
+                    SessionResultsData.getRecipientNameOfResponse(response.getId(), recipient, bundle));
+            if (isRecipientVisible) {
+                recipientTeam = recipient.getTeamName();
+            }
+        }
+
+        List<ResponseInstructorComment> responseInstructorComments =
+                bundle.getResponseCommentsMap().getOrDefault(response, Collections.emptyList());
+        List<ResponseInstructorCommentData> instructorComments =
+                SessionResultsData.buildInstructorComments(responseInstructorComments, bundle);
+        String participantComment = response.getGiverComment();
+
+        return ResponseOutput.builder()
+                .withResponseId(response.getId().toString())
+                .withGiver(giverName)
+                .withGiverUserId(giver.getGiverUserId())
+                .withGiverTeamId(giver.getTeamId())
+                .withGiverTeam(giverTeam)
+                .withGiverEmail(null)
+                .withGiverSectionName(giver.getSectionName())
+                .withGiverSectionId(giver.getSectionId())
+                .withRecipient(recipientName)
+                .withRecipientUserId(recipient.getRecipientUserId())
+                .withRecipientTeamId(recipient.getTeamId())
+                .withRecipientTeam(recipientTeam)
+                .withRecipientEmail(null)
+                .withRecipientSectionName(recipient.getSectionName())
+                .withRecipientSectionId(recipient.getSectionId())
+                .withResponseDetails(response.getFeedbackResponseDetailsCopy())
+                .withParticipantComment(participantComment)
+                .withInstructorComments(instructorComments)
+                .build();
+    }
+
+    private static String removeAnonymousHash(String identifier) {
+        return identifier.replaceAll(Const.DISPLAYED_NAME_FOR_ANONYMOUS_PARTICIPANT + " (student|instructor|team) "
+                + REGEX_ANONYMOUS_PARTICIPANT_HASH, Const.DISPLAYED_NAME_FOR_ANONYMOUS_PARTICIPANT + " $1");
+    }
+
+    public List<UserQuestionOutput> getQuestions() {
+        return questions;
+    }
+
+    /**
+     * API output format for questions in user session results.
+     */
+    public static final class UserQuestionOutput {
+
+        private final FeedbackQuestionData feedbackQuestion;
+        private final String questionStatistics;
+        private final boolean hasResponseButNotVisibleForPreview;
+        private final boolean hasCommentNotVisibleForPreview;
+
+        private final List<ResponseOutput> allResponses = new ArrayList<>();
+        private final List<ResponseOutput> responsesToSelf = new ArrayList<>();
+        private final List<ResponseOutput> responsesFromSelf = new ArrayList<>();
+        private final List<List<ResponseOutput>> otherResponses = new ArrayList<>();
+
+        private UserQuestionOutput(FeedbackQuestion feedbackQuestion, String questionStatistics,
+                boolean hasResponseButNotVisibleForPreview, boolean hasCommentNotVisibleForPreview) {
+            this.feedbackQuestion = new FeedbackQuestionData(feedbackQuestion);
+            this.questionStatistics = questionStatistics;
+            this.hasResponseButNotVisibleForPreview = hasResponseButNotVisibleForPreview;
+            this.hasCommentNotVisibleForPreview = hasCommentNotVisibleForPreview;
+        }
+
+        public FeedbackQuestionData getFeedbackQuestion() {
+            return feedbackQuestion;
+        }
+
+        public String getQuestionStatistics() {
+            return questionStatistics;
+        }
+
+        public boolean getHasResponseButNotVisibleForPreview() {
+            return hasResponseButNotVisibleForPreview;
+        }
+
+        public boolean getHasCommentNotVisibleForPreview() {
+            return hasCommentNotVisibleForPreview;
+        }
+
+        public List<ResponseOutput> getAllResponses() {
+            return allResponses;
+        }
+
+        public List<ResponseOutput> getResponsesFromSelf() {
+            return responsesFromSelf;
+        }
+
+        public List<ResponseOutput> getResponsesToSelf() {
+            return responsesToSelf;
+        }
+
+        public List<List<ResponseOutput>> getOtherResponses() {
+            return otherResponses;
+        }
+
+    }
+}
